% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termReport.R
\name{calculateTotal}
\alias{calculateTotal}
\title{calculateTotal}
\usage{
calculateTotal(roomuse, singleWeek = NA)
}
\arguments{
\item{roomuse}{RoomUse object}

\item{singleWeek}{if given roomutilization is calculated only for this week. However
number of rooms and number of lessons is calculated from entire dataset. Number of days is only
calculated for this week.}
}
\value{
fraction of periods used
}
\description{
Calculates total roomutilizition
}
\examples{
blabla
}
