% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Campus.R
\name{Campus}
\alias{Campus}
\title{Campus}
\usage{
Campus(extraLimits = "", roomgroups, lessons = list(),
  cleanupFunction = function(r) {     return(r) },
  excludeRooms = as.character(c()), comment = "", title = "")

Campus(extraLimits = "", roomgroups, lessons = list(),
  cleanupFunction = function(r) {     return(r) },
  excludeRooms = as.character(c()), comment = "", title = "")
}
\arguments{
\item{extraLimits}{sql-string (starting with AND)}

\item{roomgroups}{Array of roomgroups in WU}

\item{lessons}{list (not documentet now)}

\item{cleanupFunction}{(function that takes and return a RoomUse object)}

\item{excludeRooms}{Array}

\item{comment}{characterstring}
}
\value{
Object

S4 object of class Campus
}
\description{
Campus

Campus
}
\section{Slots}{

\describe{
\item{\code{extraLimits}}{sql-string (starting with AND)}

\item{\code{roomgroups}}{Name of roomgroup in WU}

\item{\code{lessons}}{list (not documentet now)}

\item{\code{cleanupFunction}}{(function that takes and return a RoomUse object)}

\item{\code{comment}}{characterstring}
}}

\examples{
blablabla
xxx
}
