% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countingReport.R
\name{createCountingReport}
\alias{createCountingReport}
\title{createCountingReport}
\usage{
createCountingReport(campuses, dateSelection, maxTime)
}
\arguments{
\item{campuses}{list of campuses}

\item{maxTime}{don't count minutes after maxTime}

\item{weekselection}{weekselection}
}
\value{
nothing
}
\description{
createCountingReport
}
\examples{
xxxx
}
