% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphReport.R
\name{createGraphReports}
\alias{createGraphReports}
\title{createGraphReports
Laver en grafisk rapport, som på forskellig vis illustrerer lokaleudnyttelsen.
Creates one report pr. campus, placed in defaultfolder/names.pdf (defaultname is campus_suffix)
Creates one xlsx-file with week-based counts, and one sheet pr. campus
Creates one xlsx-file with room-based counts, and one sheet pr. campus.
xlsx-files placed in defaultfolder/suffix/type_suffix.xlsx}
\usage{
createGraphReports(campuses, dateSelection, names = NULL, suffix = "x",
  ...)
}
\arguments{
\item{campuses}{En liste med et eller flere campusser (genereres med createCampus)}

\item{dateSelection}{En SQL-tekststreng, genereres med funktionen createDateSelection}

\item{names}{Et navn til hver rapport (default lokalegruppen)}

\item{suffix}{en fælles endelse til alle rapporter}

\item{...}{arguments passed to RoomUse (max- and minCapacity)}
}
\value{
nothing
}
\description{
createGraphReports
Laver en grafisk rapport, som på forskellig vis illustrerer lokaleudnyttelsen.
Creates one report pr. campus, placed in defaultfolder/names.pdf (defaultname is campus_suffix)
Creates one xlsx-file with week-based counts, and one sheet pr. campus
Creates one xlsx-file with room-based counts, and one sheet pr. campus.
xlsx-files placed in defaultfolder/suffix/type_suffix.xlsx
}
\examples{
createGraphReports(campus("AarhusN_UV"),createDateSelection(2019,5,10,maxTime=960))
}
