% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termReport.R
\name{createTermGraph}
\alias{createTermGraph}
\title{createTermGraph}
\usage{
createTermGraph(roomuse, singleWeek = NA)
}
\arguments{
\item{singleWeek}{if set, displays only that week}

\item{data}{S4 class RoomUse}
}
\value{
ggplot
}
\description{
createTermGraph
}
\examples{
none
}
