% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateSelection.R
\name{DateSelection}
\alias{DateSelection}
\title{DateSelection}
\usage{
DateSelection(year, startWeek, endWeek, maxDay = 5, maxTime = 18 * 60,
  minTime = 8 * 60, excludeWeeks = numeric(0), name = "")

DateSelection(year, startWeek, endWeek, maxDay = 5, maxTime = 18 * 60,
  minTime = 8 * 60, excludeWeeks = numeric(0), name = "")
}
\arguments{
\item{year}{numeric. starting year (2019)}

\item{startWeek}{numeric. starting week (ie. 5)}

\item{endWeek}{numeric. ending week (may be next year)}

\item{maxDay}{numeric. (default 5)}

\item{maxTime}{numeric. (minutes after midnight)}

\item{minTime}{numeric.}

\item{excludeWeeks}{numeric. (weeks to exclude from report)}

\item{name}{character name of term}
}
\value{
DateSelection

S4 object of class DateSelection
}
\description{
DateSelection

DateSelection
}
\section{Slots}{

\describe{
\item{\code{year}}{numeric. starting year (2019)}

\item{\code{startWeek}}{numeric. starting week (ie. 5)}

\item{\code{endWeek}}{numeric. ending week (may be next year)}

\item{\code{maxDay}}{numeric. (default 5)}

\item{\code{maxTime}}{numeric. (minutes after midnight)}

\item{\code{minTime}}{numeric.}

\item{\code{excludeWeeks}}{numeric. (weeks to exclude from report)}

\item{\code{name}}{name of term eg. Efterår 2019}
}}

\examples{
blablabla
blabla
}
