% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateSelection.R
\docType{methods}
\name{getSQL,DateSelection-method}
\alias{getSQL,DateSelection-method}
\title{getSQL}
\usage{
\S4method{getSQL}{DateSelection}(dateSelection)
}
\arguments{
\item{DateSelection}{}
}
\value{
a SQL-string
}
\description{
getSQL
}
\examples{
blabla
}
