% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WUroomgroups.R
\name{loadRoomGroups}
\alias{loadRoomGroups}
\title{loadRoomGroups
Load roomgroups from WebUntis. Will subsequently load rooms}
\usage{
loadRoomGroups()
}
\value{
nothing
}
\description{
loadRoomGroups
Load roomgroups from WebUntis. Will subsequently load rooms
}
\examples{
xxx
}
