% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLessonsFromWU.R
\name{loadRoomGroupsData}
\alias{loadRoomGroupsData}
\title{loadRoomGroupsData
Indæser skemabrikker for en eller flere lokalegrupper}
\usage{
loadRoomGroupsData(roomgroups, start, end, roomcount = 0,
  roomname = "")
}
\arguments{
\item{roomgroups}{En vektor bestående af navne på lokalegrupper}

\item{start}{Startdato, f.eks. "2019-01-01"}

\item{end}{Slutdato, f.eks. "2019-07-01"}

\item{roomcount}{Bruges til at begrænse antal indlæste lokaler - kun i forbindelse med tests}

\item{roomname}{begræns til et lokalenavn. Kun i forbindelse med test.}
}
\description{
loadRoomGroupsData
Indæser skemabrikker for en eller flere lokalegrupper
}
\examples{
xxx
}
