% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WUroomgroups.R
\name{loadRooms}
\alias{loadRooms}
\title{loadRooms
Load rooms from WebUntis}
\usage{
loadRooms(roomgroupsOnly = T)
}
\arguments{
\item{roomgroupsOnly}{If TRUE, only rooms belonging to a roomgroup will be loaded}
}
\value{
nothing
}
\description{
loadRooms
Load rooms from WebUntis
}
\examples{
xxx
}
