% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDefinitions.R
\name{ringetid}
\alias{ringetid}
\title{ringetid}
\usage{
ringetid(t)
}
\arguments{
\item{t}{minutter efter midnat}
}
\value{
textstreng
}
\description{
ringetid
}
\examples{
blabla
}
