% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoomUse.R
\name{RoomUse}
\alias{RoomUse}
\title{RoomUse}
\usage{
RoomUse(campus, dateSelection, minCapacity = 0, maxCapacity = 1000)

RoomUse(campus, dateSelection, minCapacity = 0, maxCapacity = 1000)
}
\arguments{
\item{campus}{S4 class campus}

\item{dateSelection}{sql-string with dateselection}

\item{minCapacity}{xx}

\item{maxCapacity}{yy}
}
\value{
RoomUse

an object of class RoomUse
}
\description{
RoomUse

RoomUse (S4-class)
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{campus}}{Campus}

\item{\code{dateSelection}}{DataSelection.}

\item{\code{dayNames}}{character.}

\item{\code{minCapacity}}{minCapacity in room}

\item{\code{maxCapacity}}{maxCapacity in room}
}}

\examples{
Use function RoomUse to create instance
blabla
}
