% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termReport.R
\name{termReport}
\alias{termReport}
\title{termReport creates in pdf-report for a campus}
\usage{
termReport(roomuse, file = NA, title = "")
}
\arguments{
\item{roomuse}{S4 class RoomUse}

\item{file}{filename}

\item{title}{title to be used in report. If "" it will be a combination of campus title and date name}
}
\value{
nothing
}
\description{
termReport creates in pdf-report for a campus
}
\examples{
blabla
}
