% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termReport.R
\name{totalsReport}
\alias{totalsReport}
\title{totalsReport}
\usage{
totalsReport(data, comment)
}
\arguments{
\item{data}{dataframe}
}
\value{
nothing
}
\description{
totalsReport
}
\examples{
blabla
}
